+++ src/lib.rs
@@ -55,6 +55,7 @@
 
 mod allocator;
 mod integration;
+mod renderer;
 
 pub use allocator::*;
 pub use integration::*;
